fix(docker): Try to enable tesing old SILE images by *not* refreshing them
Arch Linux does not support partial upgrades. We can't upgrade all the
base system packages and still expect a legacy version of SILE that we
didn't update to still run. We're going to be limited to testing against
versions we can install compatible packages for, possible from an
archive eventually.