ci: Drop tests against old SILE in Docker
We're unable to test this effectively in *these* images because Arch
does not support partial upgrades, so we can't have an up to date system
where we can install dependencies *and* hand an old SILE at the same
time. If we need regression testing farther back than the latest ICU or
whatever else caused a SILE rebuild then we'll need to build an install
SILE ourselves in a workflow, not use something premade.